requirements.txt
flask
flask-wtf
torch
torchvision
pillow
werkzeug
scipy
xformers
numpy
opencv-python